+++ packages/EventMap/CHANGELOG.md
@@ -5,8 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.3.0] - 2021-08-02
+### Added
+- Middleware
+- Logging by middleware
+### Changed
+- Converted to TypeScript
+### Removed
+- Logging by default
 
-## [1.0.0] - 2019-06-26
+
+## [0.2.0] - 2019-06-26
 ### Changed
 - Moved into sub folder with test, readme, and changelog.